+++ R/gh.R
@@ -3,7 +3,7 @@
 #' This is an extremely minimal client. You need to know the API
 #' to be able to use this client. All this function does is:
 #' * Try to substitute each listed parameter into `endpoint`, using the
-#'   `:parameter` notation.
+#'   `{parameter}` notation.
 #' * If a GET request (the default), then add all other listed parameters
 #'   as query parameters.
 #' * If not a GET request, then send the other parameters in the request
@@ -18,13 +18,14 @@
 #'
 #'    If the method is not supplied, will use `.method`, which defaults
 #'    to `"GET"`.
-#' @param ... Name-value pairs giving API parameters. Will be matched
-#'   into `endpoint` placeholders, sent as query parameters in GET
-#'   requests, and as a JSON body of POST requests. If there is only one
-#'   unnamed parameter, and it is a raw vector, then it will not be JSON
-#'   encoded, but sent as raw data, as is. This can be used for example to
-#'   add assets to releases. Named `NULL` values are silently dropped,
-#'   and named `NA` values trigger an error.
+#' @param ... Name-value pairs giving API parameters. Will be matched into
+#'   `endpoint` placeholders, sent as query parameters in GET requests, and as a
+#'   JSON body of POST requests. If there is only one unnamed parameter, and it
+#'   is a raw vector, then it will not be JSON encoded, but sent as raw data, as
+#'   is. This can be used for example to add assets to releases. Named `NULL`
+#'   values are silently dropped. For GET requests, named `NA` values trigger an
+#'   error. For other methods, named `NA` values are included in the body of the
+#'   request, as JSON `null`.
 #' @param per_page Number of items to return per page. If omitted,
 #'   will be substituted by `max(.limit, 100)` if `.limit` is set,
 #'   otherwise determined by the API (never greater than 100).
@@ -150,7 +151,6 @@ gh <- function(endpoint, ..., per_page = NULL, .token = NULL, .destfile = NULL,
 
   params <- list(...)
   params <- drop_named_nulls(params)
-  check_named_nas(params)
 
   if (is.null(per_page)) {
     if (!is.null(.limit)) {
@@ -167,6 +167,9 @@ gh <- function(endpoint, ..., per_page = NULL, .token = NULL, .destfile = NULL,
                           overwrite = .overwrite, accept = .accept,
                           send_headers = .send_headers,
                           api_url = .api_url, method = .method)
+
+
+  if (req$method == "GET") check_named_nas(params)
 
   if (.progress) prbr <- make_progress_bar(req)