+++ R/gh_request.R
@@ -63,10 +63,14 @@ gh_set_endpoint <- function(x) {
       value    = params[[i]][1],
       template = endpoint
     )
-    if (endpoint2 != endpoint) {
-      endpoint <- endpoint2
-      done[i] <- TRUE
+    if (is.na(endpoint2)) {
+      throw(new_error(
+        "Named NA parameters are not allowed: ", names(params)[i],
+        call. = FALSE
+      ))
     }
+    endpoint <- endpoint2
+    done[i] <- TRUE
     if (!is_template(endpoint)) {
       break
     }